+++ src/bp/core/botpress.ts
@@ -1,8 +1,8 @@
 import * as sdk from 'botpress/sdk'
 import lang from 'common/lang'
+import { addStepToEvent, EventCollector, StepScopes, StepStatus } from 'core/events/event-collector'
 import { EventEngine } from 'core/events/event-engine'
-import { LogsJanitor } from 'core/logger/logs-janitor'
-
+import { LoggerDbPersister, LoggerFilePersister, LoggerProvider, LogsJanitor } from 'core/logger'
 import { copyDir } from 'core/misc/pkg-fs'
 import { WrapErrorsWith } from 'errors'
 import fse from 'fs-extra'
@@ -23,8 +23,6 @@ import { createForGlobalHooks } from './api'
 import { BotpressConfig } from './config/botpress.config'
 import { ConfigProvider } from './config/config-loader'
 import Database from './database'
-import { addStepToEvent, EventCollector, StepScopes, StepStatus } from './events/event-collector'
-import { LoggerDbPersister, LoggerFilePersister, LoggerProvider } from './logger'
 import { ModuleLoader } from './module-loader'
 import { WellKnownFlags } from './sdk/enums'
 import { Event } from './sdk/impl'
@@ -43,7 +41,6 @@ import { DialogJanitor } from './services/dialog/janitor'
 import { SessionIdFactory } from './services/dialog/session/id-factory'
 import { HintsService } from './services/hints'
 import { Hooks, HookService } from './services/hook/hook-service'
-
 import { StateManager } from './services/middleware/state-manager'
 import { MigrationService } from './services/migration'
 import { MonitoringService } from './services/monitoring'
