+++ src/bp/core/logger/index.ts
@@ -2,3 +2,5 @@ export * from './persisted-console-logger'
 export * from './logger-db-persister'
 export * from './logger-file-persister'
 export * from './logs-service'
+export * from './logs-janitor'
+export * from './logs-repository'
