+++ src/bp/core/routers/admin/index.ts
@@ -3,7 +3,7 @@ import { Logger } from 'botpress/sdk'
 import { checkRule } from 'common/auth'
 import LicensingService from 'common/licensing-service'
 import { ConfigProvider } from 'core/config/config-loader'
-import { LogsRepository } from 'core/logger/logs-repository'
+import { LogsRepository } from 'core/logger'
 import { ModuleLoader } from 'core/module-loader'
 import { GhostService } from 'core/services'
 import { AlertingService } from 'core/services/alerting-service'
