+++ src/bp/core/services/bot-monitoring-service.ts
@@ -1,7 +1,7 @@
 import 'bluebird-global'
 import { Logger, LoggerEntry } from 'botpress/sdk'
 import { ConfigProvider } from 'core/config/config-loader'
-import { LogsRepository } from 'core/logger/logs-repository'
+import { LogsRepository } from 'core/logger'
 import { TYPES } from 'core/types'
 import { inject, injectable, tagged } from 'inversify'
 import { Redis } from 'ioredis'
