+++ src/bp/core/services/services.inversify.ts
@@ -4,8 +4,7 @@ import { EventEngine } from 'core/events/event-engine'
 import { Queue } from 'core/events/queue'
 import MemoryQueue from 'core/events/queue/memory-queue'
 import { KeyValueStore } from 'core/kvs'
-import { LogsJanitor } from 'core/logger/logs-janitor'
-import { LogsService } from 'core/logger/logs-service'
+import { LogsJanitor, LogsService } from 'core/logger'
 import { DialogContainerModule } from 'core/services/dialog/dialog.inversify'
 import { CEJobService, JobService } from 'core/services/job-service'
 import { ContainerModule, interfaces } from 'inversify'
